% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lda.R
\name{.cosine_matrix}
\alias{.cosine_matrix}
\title{Cosine similarity matrix among rows of a matrix}
\usage{
.cosine_matrix(m)
}
\description{
Cosine similarity matrix among rows of a matrix
}
\keyword{internal}
