% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.entropy}
\alias{.entropy}
\title{Shannon entropy of a proportion vector (log2)}
\usage{
.entropy(p)
}
\arguments{
\item{p}{Numeric vector of proportions}
}
\value{
Scalar entropy in bits
}
\description{
Shannon entropy of a proportion vector (log2)
}
\keyword{internal}
