% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{.get_membership}
\alias{.get_membership}
\title{Retrieve a fitted model's membership object from an \code{sp} object}
\usage{
.get_membership(sp, model)
}
\arguments{
\item{sp}{A \code{symbayes} object}

\item{model}{One of "dmm", "lda", "hdp"}
}
\value{
The membership object
}
\description{
Retrieve a fitted model's membership object from an \code{sp} object
}
\keyword{internal}
