% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-hdp.R
\name{.hdp_build}
\alias{.hdp_build}
\title{Build the HDP tree structure (flat or grouped)}
\usage{
.hdp_build(sp, mat, group_var, alphaa, alphab, n_categ, n_samp)
}
\description{
Build the HDP tree structure (flat or grouped)
}
\keyword{internal}
