% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lda.R
\name{.lda_search}
\alias{.lda_search}
\title{LDA K-search over a range, scoring redundancy and pure samples}
\usage{
.lda_search(dtm, k_range, redundant_thresh, control)
}
\description{
LDA K-search over a range, scoring redundancy and pure samples
}
\keyword{internal}
