% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dmm.R
\name{.map_symportal}
\alias{.map_symportal}
\title{Map SymPortal dominant profile and clade onto metadata}
\usage{
.map_symportal(sp, md, model)
}
\description{
Map SymPortal dominant profile and clade onto metadata
}
\keyword{internal}
