% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{print.symbayes}
\alias{print.symbayes}
\title{Print method for symbayes objects}
\usage{
\method{print}{symbayes}(x, ...)
}
\arguments{
\item{x}{A \code{symbayes} object}

\item{...}{Ignored}
}
\description{
Print method for symbayes objects
}
