% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{print.symbayes_membership}
\alias{print.symbayes_membership}
\title{Print method for membership objects}
\usage{
\method{print}{symbayes_membership}(x, ...)
}
\arguments{
\item{x}{A \code{symbayes_membership} object}

\item{...}{Ignored}
}
\description{
Print method for membership objects
}
